Временно сделаем такое же имя модуля, как для реакта, чтобы не сломались завязки на имя https://online.sbis.ru/opendoc.html?guid=60bd4acd-32c2-4fdb-936f-10d5c20ecc58
