+++ UIInferno/ReactUnitTest/Focus/ConditionContent.ts
@@ -2,10 +2,11 @@ import { TemplateFunction } from 'UICommon/Base';
 import {TestBaseControl} from '../Base';
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/MinusOneTabindex');
+import template = require('wml!ReactUnitTest/Focus/ConditionContent');
 
 class TestControl extends TestBaseControl {
-   _template: TemplateFunction = template;
+    _template: TemplateFunction = template;
+    noNeedContent: boolean = false;
 }
 
 export default TestControl;
