+++ UIInferno/ReactUnitTest/Focus/ContentEditableDiv.ts
@@ -2,7 +2,7 @@ import { TemplateFunction } from 'UICommon/Base';
 import {TestBaseControl} from '../Base';
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/Input');
+import template = require('wml!ReactUnitTest/Focus/ContentEditableDiv');
 
 class TestControl extends TestBaseControl {
    _template: TemplateFunction = template;
