+++ UIInferno/ReactUnitTest/Focus/ContentEditableDivWithLink.ts
@@ -2,7 +2,7 @@ import { TemplateFunction } from 'UICommon/Base';
 import {TestBaseControl} from '../Base';
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/ContentEditableDiv');
+import template = require('wml!ReactUnitTest/Focus/ContentEditableDivWithLink');
 
 class TestControl extends TestBaseControl {
    _template: TemplateFunction = template;
