+++ UIInferno/ReactUnitTest/Focus/DelegatedTabfocus.ts
@@ -1,8 +1,8 @@
 import { TemplateFunction } from 'UICommon/Base';
-import {TestBaseControl} from '../Base';
+import {TestBaseControl} from '../Base'
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/WithDefaultOpener');
+import template = require('wml!ReactUnitTest/Focus/DelegatedTabfocus');
 
 class TestControl extends TestBaseControl {
     _template: TemplateFunction = template;
