+++ UIInferno/ReactUnitTest/Focus/Input.ts
@@ -1,0 +1,11 @@
+import { TemplateFunction } from 'UICommon/Base';
+import {TestBaseControl} from '../Base';
+
+// @ts-ignore
+import template = require('wml!ReactUnitTest/Focus/Input');
+
+class TestControl extends TestBaseControl {
+   _template: TemplateFunction = template;
+}
+
+export default TestControl;
