+++ UIInferno/ReactUnitTest/Focus/MinusOneTabindex.ts
@@ -2,7 +2,7 @@ import { TemplateFunction } from 'UICommon/Base';
 import {TestBaseControl} from '../Base';
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/AutofocusInside');
+import template = require('wml!ReactUnitTest/Focus/MinusOneTabindex');
 
 class TestControl extends TestBaseControl {
    _template: TemplateFunction = template;
