+++ UIInferno/ReactUnitTest/Focus/Proxy.ts
@@ -2,7 +2,7 @@ import {Control} from 'UICore/Base';
 import { TemplateFunction } from 'UICommon/Base';
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/Proxy');
+import template = require('wml!ReactUnitTest/Focus/Proxy');
 
 class TestControl extends Control {
     _template: TemplateFunction = template;
