+++ UIInferno/ReactUnitTest/Focus/Simple.ts
@@ -2,7 +2,7 @@ import { TemplateFunction } from 'UICommon/Base';
 import {TestBaseControl} from '../Base';
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/ContentEditableDivWithLink');
+import template = require('wml!ReactUnitTest/Focus/Simple');
 
 class TestControl extends TestBaseControl {
    _template: TemplateFunction = template;
