+++ UIInferno/ReactUnitTest/Focus/SvgWithNoFocus.ts
@@ -2,7 +2,7 @@ import { TemplateFunction } from 'UICommon/Base';
 import {TestBaseControl} from '../Base';
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/SvgWithNoFocus');
+import template = require('wml!ReactUnitTest/Focus/SvgWithNoFocus');
 
 class TestControl extends TestBaseControl {
     _template: TemplateFunction = template;
