+++ UIInferno/ReactUnitTest/Focus/SvgWithNoFocus.wml
