+++ UIInferno/ReactUnitTest/Focus/WithDefaultOpener.ts
@@ -2,10 +2,10 @@ import { TemplateFunction } from 'UICommon/Base';
 import {TestBaseControl} from '../Base';
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/Simple');
+import template = require('wml!ReactUnitTest/Focus/WithDefaultOpener');
 
 class TestControl extends TestBaseControl {
-   _template: TemplateFunction = template;
+    _template: TemplateFunction = template;
 }
 
 export default TestControl;
