+++ UIInferno/ReactUnitTest/Focus/WithInputFakeMobile.ts
@@ -3,7 +3,7 @@ import {TestBaseControl} from '../Base';
 import {detection} from 'Env/Env';
 
 // @ts-ignore
-import template = require('wml!UICoreTest/Focus/WithInputFakeMobile');
+import template = require('wml!ReactUnitTest/Focus/WithInputFakeMobile');
 
 class TestControl extends TestBaseControl {
     _template: TemplateFunction = template;
