+++ UIInferno/ReactUnitTest/ReactUnitTest.s3mod
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="UTF-8" ?>
-<ui_module id="62bf645b-207a-4e83-b06d-41e8c51d6485" name="UICoreTest" version="1.00">
+<ui_module id="62bf645b-207a-4e83-b06d-41e8c51d6485" name="ReactUnitTest" version="1.00">
 
   <description/>
