+++ UIInferno/ReactUnitTest/_vdom/Synchronizer/resources/VdomMarkup.test.js
