+++ UIInferno/ReactUnitTest/_vdom/Utils/Functional.test.js
