--- UIInferno/UICoreTest/Focus/AutofocusInside.wml
@@ -1,4 +1,0 @@
-<div>
-   Can focus it
-   <UICoreTest.Focus.Simple name="autofocusChild" ws-autofocus="true" />
-</div>
